Single season block bootstrap prediction intervals through time series cross-validation forecasting

Compute prediction intervals by applying the single season block bootstrap method to subsets of time series data using a rolling forecast origin.

Usage

bb_cvforecast(
  object,
  data,
  yvar,
  neighbour = 0,
  predictor.vars,
  h = 1,
  season.period = 1,
  m = 1,
  num.futures = 1000,
  level = c(80, 95),
  forward = TRUE,
  initial = 1,
  window = NULL,
  roll.length = 1,
  recursive = FALSE,
  recursive_colNames = NULL,
  ...
)

Arguments

object

Fitted model object of class smimodel, backward, gaimFit or pprFit.

data

Data set. Must be a data set of class tsibble.(Make sure there are no additional date or time related variables except for the index of the tsibble). If multiple models are fitted, the grouping variable should be the key of the tsibble. If a key is not specified, a dummy key with only one level will be created.

yvar

Name of the response variable as a character string.

neighbour

If multiple models are fitted: Number of neighbours of each key (i.e. grouping variable) to be considered in model fitting to handle smoothing over the key. Should be an integer. If neighbour = x, x number of keys before the key of interest and x number of keys after the key of interest are grouped together for model fitting. The default is neighbour = 0 (i.e. no neighbours are considered for model fitting).

predictor.vars

A character vector of names of the predictor variables.

h

Forecast horizon.

season.period

Length of the seasonal period.

m

Multiplier. (Block size = NULLseason.period * m)

num.futures

Number of possible future sample paths to be generated.

level

Confidence level for prediction intervals.

forward

If TRUE, the final forecast origin for forecasting is \(y_T\). Otherwise, the final forecast origin is \(y_{T-1}\).

initial

Initial period of the time series where no cross-validation forecasting is performed.

window

Length of the rolling window. If NULL, a rolling window will not be used.

roll.length

Number of observations by which each rolling/expanding window should be rolled forward.

recursive

Whether to obtain recursive forecasts or not (default - FALSE).

recursive_colNames

If recursive = TRUE, a character vector giving the names of the columns in test data to be filled with forecasts. Recursive/autoregressive forecasting is required when the lags of the response variable itself are used as predictor variables into the model. Make sure such lagged variables are positioned together in increasing lag order (i.e. lag_1, lag_2, ..., lag_m, lag_m = maximum lag used) in data, with no break in the lagged variable sequence even if some of the intermediate lags are not used as predictors.

...

Other arguments not currently used.

See also