Converting a fitted gam object to a smimodelFit object

Converts a given object of class gam to an object of class smimodelFit.

Usage

make_smimodelFit(
  x,
  yvar,
  neighbour,
  index.vars,
  index.ind,
  index.data,
  index.names,
  alpha,
  s.vars = NULL,
  linear.vars = NULL
)

Arguments

x

A fitted gam object.

yvar

Name of the response variable as a character string.

neighbour

If multiple models are fitted: Number of neighbours of each key (i.e. grouping variable) to be considered in model fitting to handle smoothing over the key. Should be an integer. If neighbour = x, x number of keys before the key of interest and x number of keys after the key of interest are grouped together for model fitting. The default is neighbour = 0 (i.e. no neighbours are considered for model fitting).

index.vars

A character vector of names of the predictor variables for which indices are estimated.

index.ind

An integer vector that assigns group index for each predictor in index.vars.

index.data

A tibble including columns for the constructed indices.

index.names

A character vector of names of the constructed indices.

alpha

A vector of index coefficients.

s.vars

A character vector of names of the predictor variables for which splines are fitted individually (rather than considering as part of an index).

linear.vars

A character vector of names of the predictor variables that are included linearly in the model.

Value

An object of class smimodelFit, which is a list that contains following elements:

alpha

A sparse matrix of index coefficients vectors. Each column of the matrix corresponds to the index coefficient vector of each index.

derivatives

A tibble of derivatives of the estimated smooths.

var_y

Name of the response variable.

vars_index

A character vector of names of the predictor variables for which indices are estimated.

vars_s

A character vector of names of the predictor variables for which splines are fitted individually.

vars_linear

A character vector of names of the predictor variables that are included linearly in the model.

neighbour

Number of neighbours of each key considered in model fitting.

gam

Fitted gam.