Possible future sample paths (multi-step) from smimodel residuals

Generates possible future sample paths (multi-step) using residuals of a fitted smimodel through recursive forecasting.

Usage

possibleFutures_smimodel(object, newdata, bootstraps, recursive_colRange)

Arguments

object

A smimodel object.

newdata

The set of new data on for which the forecasts are required (i.e. test set; should be a tsibble).

bootstraps

Generated matrix of bootstrapped residual series.

recursive_colRange

The range of column numbers in newdata to be filled with forecasts.