smimodel: Sparse Multiple Index Models for Nonparametric Forecasting

Implements a general algorithm for estimating Sparse Multiple Index (SMI) models for nonparametric forecasting and prediction. Includes functions for fitting nonparametric additive models with backward elimination, group-wise additive index models, and projection pursuit regression models as benchmark comparison methods. Provides tools for generating prediction intervals to quantify uncertainty in point forecasts produced by the SMI model and benchmark models, using the classical block bootstrap method and a new method called conformal bootstrap, which integrates block bootstrap with split conformal prediction.

See also

Author

Maintainer: Nuwani Palihawadana Nuwani.KodikaraPalihawadana@monash.edu (ORCID) [copyright holder]

Other contributors:

Xiaoqian Wang Xiaoqian.Wang@amss.ac.cn (ORCID) [contributor]